Add more documentation
